bunch of small stuff
Fixed the issue when going to a game page from mylist and the game has similarly named other games. Bayonetta
would redirect to 8-bit Bayonetta instead, it'll no longer do this.

Fixed the issue when going to a game page and the "unauthorized token!" toast would be popping up even when
your token was still valid

changed the color of the responsive bars/cross icons that shows up when shrinking the window to white

invisible table delete bug towards bottom of the mylist list made a bit better

fixed toast positioning

upon registering a toast pops up telling you that you've registered and that an email verification was sent

upon requesting a password reset a toast pops up telling you that a password email was sent

steam game add alerts and every alert in general that I could find has been removed, a toast will now say whether or
not the steam game add was successful